+++ src/main/java/org/zeros/farm_manager_server/Domain/DTO/Crop/CropParameters/SugarBeetParametersDTO.java
@@ -3,6 +3,8 @@ package org.zeros.farm_manager_server.Domain.DTO.Crop.CropParameters;
 import lombok.*;
 import lombok.experimental.SuperBuilder;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @AllArgsConstructor
@@ -10,5 +12,5 @@ import lombok.experimental.SuperBuilder;
 @EqualsAndHashCode(callSuper = true)
 @SuperBuilder
 public class SugarBeetParametersDTO extends CropParametersDTO {
-    private float sugarContent;
+    private BigDecimal sugarContent;
 }
